Reland "[views-ax] Migrate StringAttribute::kDescription in ui/views/* and chrome/browser/*"
This is a reland of commit e2b7babe257020988e3130d6a4719af774430a2d

Failure happened in chrome build:
https://issues.chromium.org/issues/358594668

Reason for revert:
The failure happened because the browser test
PasswordGenerationPopupViewTest.HeaderAccessibilityProperties incorporated
a hard coded string (IDS_PASSWORD_GENERATION_HELP_TEXT_TRUSTED_ADVICE)
to compute the expected accessible description.
On careful inspection, it was discovered that depending on some
features being enabled the accessible description changes and as such,
there was a need to correct the expected accessible description in the browser test itself.

The fix includes correctly (or dynamically) computing the message ids required to obtain the
expected accessible description in the test, so no matter what features are enabled
the expected accessible description and description in the cache are always same.

Original change's description:
> [views-ax] Migrate StringAttribute::kDescription in ui/views/* and chrome/browser/*
>
> This CL migrates the kDescription attribute in Views to be updated
> whenever its value should change, rather than querying the value
> and computing it only when needed.
>
> This CL is part of the ViewsAX project:
> https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413
>
> Bug: 325137417
> Change-Id: I156378736ef05badca157ecb4073322f8f631935
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5744687
> Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
> Reviewed-by: Javier Contreras <javiercon@microsoft.com>
> Reviewed-by: Scott Violet <sky@chromium.org>
> Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
> Cr-Commit-Position: refs/heads/main@{#1339830}

Bug: 325137417
Bug: 358594668
Change-Id: If0e96e8900eb29619e775a9fc49a0cc6092c0878
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5781349
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Divyansh Mangal <dmangal@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1341116}