[iOS] Add OmahaServiceObserver to Improve App Update Handling
This CL introduces OmahaServiceObserver to improve the reliability and
safety of app updates on iOS. The observer pattern is leveraged to
promptly notify clients of available updates and any errors encountered
during the process. IOSChromeSafetyCheckManager is also updated to act
as an OmahaServiceObserver, enhancing the app's update management
capabilities. Additionally, the kOmahaServiceRefactor killswitch is
introduced to enable a gradual and safe migration of the Omaha Service.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED

Change-Id: I1d036631b708c2a8b8c2f0bc4f59e1550b781ae0
Fixed: 347975316
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5597548
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341121}