[IOSSignout] Make the Legacy Data Clearing logic apply before new one
The logic to clear all data regardless of the time of sign in should
take precedence. This is to handle the case when a user signs in with
sync and all the data is stored in local storage. Once sync-to-signin
feature is enabled while signed in, the old data will stay in local
storage.cThe old data synced to local storage will remain on the device
if not all data is cleared.

The CL also improves tests to make a distinction between the two
`ClearBrowsingData*` calls.

Bug: 358564921
Change-Id: I53e2bef1036129adac5381d4678055ce73eae6c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5777067
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1341135}