ATSPI Fuzzer: cache control paths
This fuzzer fuzzes the accessibility tree by interacting with controls.
The key thing about that tree is that it's a tree; each control has a
variety of child controls.

Previously, this fuzzer:
* Recorded control names and roles during fuzzing; then
* Used those recorded names and roles during mutation.

Despite a lot of tinkering, this has been ineffective at reaching into
deep controls. All sorts of heuristics have been tried but ultimately
the odds of inserting a valid child into a fuzz case seem to be just too
low.

So, this change ceases to cache control names and roles but instead
caches entire control paths. Specifically,
* When we're fuzzing and poking at a given control, we record the
  entire path to each of the child controls
* When we're mutating fuzz cases, we may entirely replace a control path
  with some other known-valid control path.

In local testing this seems much more effective at reaching into
controls deep inside the hierarchy.

This is a compatibility break with existing fuzz cases because it adds
an extra layer of indirection into the proto. This is no big deal
because the fuzzer hasn't been too effective so far.

As part of the change, this also:
* Includes code to keep the cache database size reasonable.
* Removes code which skipped over the first few layers of controls,
  since it didn't have the desired effect of helping the algorithm get
  deeper, and made the test cases more confusing. Instead we explicitly
  treat the first-level node as anonymous (since its name varies based
  on RAM usage).

Cq-Include-Trybots: luci.chromium.try:linux-centipede-asan-rel,win-libfuzzer-asan-rel,linux-libfuzzer-asan-rel
Change-Id: I13c0418142ebff9e4b4c1a9c3bffbda4fabab8b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782825
Reviewed-by: Paul Semel <paulsemel@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341167}