Fix potential reentrancy in WebXr permissions flow
WebXr triggers additional permissions requests for features only if the
general permission for the session was granted. This is done to ensure
that the user only sees the prompts for additional features (which may
be optional) if they agree to the overall immersive experience.

When requesting a permission, the permissions infrastructure determines
if the new permission should "preempt" the current permission request.
In some cases for WebXr, this "feature" level permission is chosen
to preempt the overall permission. This causes some re-entrant behavior,
as the list of permissions currently being iterated over to resolve the
user's response is cleared in preparation for requesting the new
permission. When flow returns to finish the user's response to the first
permission, the iterator being used by the loop is now invalid, and
errors occur.

While this likely should be fixed within the permissions code, WebXr is
one of the few places that has such a complex flow as to later prompt
for additional features ourselves in this decision callback. This change
thus patches the WebXr code until a decision about a proper and more
thorough fix can be decided upon within permissions code, in order to
unblock ongoing work for features that require a new permission.

Bug: 357776212
Change-Id: I785613eeba3d6f08feb9111a748f469e1cb74068
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785799
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341173}