[Desktop Android CRX] Compile code in //extensions/renderer
We are experimenting with desktop-android configurations for Chrome.

This CL includes the code in //extensions/renderer in the desktop-
android build configurations. This directory includes code for injecting
extension JS bindings for a given context and setting up access to APIs.

This just includes the code in the build, and does not provide any way
for it to be reached, nor any way to trigger the load of an extension.

This is currently experimental and has no production behavior changes.

Cq-Include-Trybots: luci.chromium.try:android-desktop-arm64-compile-rel,android-desktop-x64-compile-rel

Bug: 356905053
Change-Id: I5dc6c4f6616ae4a9c62fc0b08abffa1035ad269a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5774138
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Auto-Submit: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341222}