Ensure correct rounded corner bounds for server-side rounded windows
Currently, when the windows have square rounded corners, we still set
rectangular rounded corner bounds. Visually, the windows will be square
but there is a performance penalty for setting square MaskFilterInfo
for a quad.

By setting non-empty MaskFilterInfo, we enable blending for the
quad, which is expensive. [1]

[1] https://source.chromium.org/chromium/chromium/src/+/main:components/viz/common/quads/draw_quad.h;l=95-98;drc=5203366d4cf174b9331ee33563e0520f10b60828

Bug: b/345488202
Change-Id: I55ef88440886119714f7f0830873d794c0ecdf95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5784008
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341245}