Fix Stale While Revalidate tests in chrome
The tests were passing in content_shell runner but were failing in
chrome. The problem was that the simple disk cache was tagging the entry
as unusable due to it having a freshness of 0. The entry is still useful
if the staleness is non-zero as well.

Fixes wpt/fetch/stale-while-revalidate/* in the browser.

Change-Id: Ic9738ff6f01d7fae05594f3708b138f0648e803d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5786419
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341306}