WebNN: avoid read back buffers after dispatch
Fixes an issue when readBuffer() gets called right after dispatch(),
which whips up a staging buffer because it requires the buffer to
be mapped, which may not be allowed if the GPU was still using it. Instead, read buffer now enqueues map, which avoids falling back to staging while dispatch() is still executing.

Bug: 358594662
Change-Id: Ic29c4d92c8b704803cbfa31a5b4ccc8025c46ff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782586
Reviewed-by: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Commit-Queue: Bryan Bernhart <bryan.bernhart@intel.com>
Cr-Commit-Position: refs/heads/main@{#1341316}