Update HTTPS-First Mode settings metrics
With the addition of a second backing preference for Balanced Mode, the
Security.HttpsFirstMode.SettingChanged histogram would be emitted twice
because the pref observer would fire twice every time the UI setting was
changed, double counting the metric. This moves the "setting changed"
collection to the GeneratedHttpsFirstModePref could so that it is
logged at the point where we convert the new UI setting choice into the
underlying backing prefs, which is only fired once each time the UI
setting changes.

This also simplifies the startup-setting metric collection be the same
regardless of the Balanced Mode feature flag, which simplifies the
collection and testing. (Distinguishing these cases isn't worth the
additional complexity here, so it's nicer to just fully move over to the
new histogram.)

Bug: 349860796
Change-Id: Icb1f650901a7c6fc0816d61e42bdac9b467c3fb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785049
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341457}