2nd speculative fix for flaky PointeroverOrPointerdownWillWarmUpServiceWorker test
The previous speculative fix did not work unfortunately. This is yet
another speculative fix. The assumption is that moving to (1000, 1000)
might fail on some platform. So this CL changes the location to (0, 0),
and also creates out_scope_url link, and moves the mouse pointer to
out_scope_url link first, and move to in_scope_url link.

Bug: 349700657
Change-Id: Id1ce38629f0df0a1c0c945a558e11c2075ca604b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785280
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341470}