Clear visual properties when setting the main frame routing ID
Prior to this change, a race condition existed when a WebViewImpl
transitions from being a remote frame to a local frame. The updated
visual properties were not sent based on the timing of when different
navigations completed.

This race condition was commonly seen in Microsoft Edge with an internal
site which kicked off a top-level nav to another origin and immediately
afterward navigated an iframe to a page of the same origin.

The included test emulates this situation. Prior to this change, the
test consistently fails with the renderer having a 0x0-sized viewport.

Bug: 352093463
Change-Id: Ia2bab87a040395a534fda21069250b5fd08df6e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785294
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Auto-Submit: Erik Anderson <Erik.Anderson@microsoft.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1341485}