[freetype] Reduce max hint count when fuzzing
It is possible to create an infinite loop in glyph hinting instructions.
FreeType protects against this by limiting to 1,000,000 total
instructions executed by default. However, this can still take a long
time depending on the instructions used. Instead, limit the number of
instructions to execute to 10,000 when building for a fuzzer. This
is enough to test for interesting interactions (and is more than most
fonts use) but low enough to speed up fuzzing to find useful sets of
instructions instead of running infinite loops.

Bug: chromium:356637049
Change-Id: I1ba0b222c183bf84c94cb62147c4b29c5314f132
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783364
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341493}