drivefs_search: Run search callback with ABORT if disconnected / unbound
This guarantees that the search callback is always run, and allows us to
remove the `mojo::WrapCallbackWithDefaultInvokeIfNotRun` wrappers from
all callers of `DriveFsHost::PerformSearch` is called.

Note that this CL calls `DriveFsSearchQuery::OnGetNextPage` from
`DriveFsSearchQuery::GetNextPage` when the remote is not bound, or when
the callback passed into the `mojom::SearchQuery::GetNextPage` is
destructed. This is equivalent to calling the callback directly instead
of going through `OnGetNextPage`, but is done in this manner:

- to ensure that multiple `mojo::WrapCallbackWithDefaultInvokeIfNotRun`
  wrappers are not nested when `DriveFsSearchQuery::GetNextPage` is
  called again from `OnGetNextPage`, and
- in case `OnGetNextPage` is refactored later to deal with this
  `drive::FILE_ERROR_ABORT` case.

Bug: b:357980197
Change-Id: I6a6a2100f5ec4e4734d38e3ef02d8e5dc8be2420
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5781250
Commit-Queue: Michael Cui <mlcui@google.com>
Reviewed-by: CJ Huang <chenjih@google.com>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341501}