[BatchUpload] Create core structures for the Batch Upload feature
batch_upload: contains the main interface for client usages, also contains internal group of functions that allow easily create the
information needed by the BatchUploadController without it needing to
be aware of all the data types details.

batch_upload_data_provider: contains the interface for data types that
wants to integrate with the Batch Upload dialog, and the structures
that goes along with it.

batch_upload_controller: Central structure that connects the different
data types to the UI component by communicating with the different
types to get the local data to show, as well as triggering the UI and
listening to the responses to finally forward back to each data type
to perform the move to the account storage operations.

Added tests structures with minimal tests as a start, mostly through
fake implementations.

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL The CL is mainly to create the structure, most of the lines are tested with dummy values for now, the real tests will be added in follow up CLs.
Bug: b:358553569
Change-Id: I6456240f7def89396fd63194dc0d7348a5f77feb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5767300
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1341543}