Prefetch: Call `PreloadingAttemptImpl::SetIsAccurateTriggering()` unconditionally
This CL moves a call of
`PreloadingAttemptImpl::SetIsAccurateTriggering()` from
`PrefetchService::OnMaybeDeterminedHead()` to
`PrefetchContainer::OnReturnPrefetchToServe()` to minimize a diff to
https://crrev.com/c/5670092.

There are two paths and cases that
`PreloadingAttemptImpl::SetIsAccurateTriggering()` is called and true is
set.

A. When navigation started. From
   `PreloadingDataImpl::DidStartNavigation()`. True is set iff the
   prefetch `IsExactlyMatch()` or `IsNoVarySearchHeaderMatch()`.
   Note that `IsNoVarySearchHeaderMatch()` is false if the prefetch is
   `kShouldBlockUntilHead` state.
B. When match end timing with some condition. (See below.) True is set
   iff the prefetch `IsExactlyMatch()` or
   `IsNoVarySearchHeaderMatch()`. Note that it is ensured that
   the prefetch is not `kShouldBlockUntilHead` state.

Reg. B. Before this CL,
`PreloadingAttemptImpl::SetIsAccurateTriggering()` was called in
`PrefetchService::OnMaybeDeterminedHead()`, which was called for
prefetch only if it was `kBlockUntilHead` state at the match start
timing. After this CL, instead, we call it when (approximately) match
end, for a prefetch that was potentially matching to a navigation and
not `kNotServable` at the match start timing, without additional
condition.

Note that this CL reduces the call for prefetches that are
`PrefetchContainer::HasPrefetchBeenConsideredToServe()`, because the
path [1] doesn't call `PrefetchContainer::OnReturnPrefetchToServe()`.
This doesn't change metrics because `AccurateTriggering` is already set
true when this is served first time.

Behavioral changes:

- Prefetch can be used in navigations that share initiator document,
  e.g. prerenders. Consider the following scenario: The primary page A
  starts prefetch B and prerender C. C can use B, but it was ignored in
  `PreloadingDataImpl::DidStartNavigation()`. So, this CL changes
  `AccurateTriggering` to be true for this case.
- Other edge cases for B is unknown so far.

More improvements will be done under a feature flag
`kPrefetchNewWaitLoop` [2] instead of investigating edge cases.

Note that this CL changes behavior of unit tests as A is not triggered
in unit tests.

[1] https://source.chromium.org/chromium/chromium/src/+/refs/tags/129.0.6614.0:content/browser/preloading/prefetch/prefetch_service.cc;l=1522
[2] https://chromium-review.googlesource.com/c/chromium/src/+/5670092

Bug: 353489607, 343867153
Change-Id: Ifc72dced41c03bea37f88b1fac8c4dd18c41af97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5764864
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Ken Okada <kenoss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341575}