[Android] Show correct KA action title when passkeys are available.
Before this CL, only the WebAuthnCredentialsDelegate is checked for
passkeys availability. The only functional change of this CL is that
WebAuthnCredManDelegate is also checked if it has passkeys available
for display. If so, the "Manage passwords and passkeys" UI string is
used instead of the "Manage passwords" for the KA management action.

Fixed: 359768395
Change-Id: Ibdf15e4f7dcd6719d61158362acb8b28440e922c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783823
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1341586}